+++ src/utils.ts
@@ -31,7 +31,7 @@ export const EMPTY_FUNCTION: () => void = Object.freeze(() => {
  */
 export const DEFAULT_WORKER_CHOICE_STRATEGY_OPTIONS: WorkerChoiceStrategyOptions =
   {
-    choiceRetries: 6,
+    retries: 6,
     runTime: { median: false },
     waitTime: { median: false },
     elu: { median: false }